+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Initial project created
 - Standard repo files added
+- Parse command arguments